+++ erdpy/projects/project_clang.py
@@ -13,7 +13,7 @@ logger = logging.getLogger('ProjectClang')
 
 class ProjectClang(Project):
 
-    def __init__(self, directory):
+    def __init__(self, directory: Path):
         super().__init__(directory)
         self.ensure_config_file()
 
@@ -99,7 +99,7 @@ class ProjectClang(Project):
 
         myprocess.run_process(args)
 
-    def _do_after_build(self) -> List[Path]:
+    def _do_after_build_custom(self) -> List[Path]:
         output_wasm_file = self._copy_to_output(self.file_output)
         self.file_output.unlink()
         self.file_ll.unlink()